Deleted password configuration and small changes
I thought several times if it's really necessary to activate a password to this program, and finally, I realized that it doesn't seem very important.
Instead, I decided to create a new field named "username" which I'll use to welcome the user. So basically I added "username configuration" fields and some new messages in the "Languages" class.