Basically, a big change
Basically, a big change.

I wrote some new lines of code related to the ConfigurationController class (in order to work with the new function Component.runSwitchButtonEffect) and some minor changes in Language and Colour classes.

The main changes are inside ConfigurationPanel and Component classes. I've finally given use to the new Component class functions in ConfigurationPanel, removing A LOT of code in the process.
I must say, it is more readable.