Added some new cool stuff!
Summarizing, I added a customized JOptionPane and added in a new class named "Advice". This was the last thing I wanted to do that didn't succeed, until now! (Believe it or not, it will make me work less in the future).
Also, I added some new text in the Language class (e.g. the software title) and I added an ImageIcon as well!