Finally, another big step!
I created some new classes, but the most important definitively is "MainController" class. Now, I can successfully change the current panel on the frame. Because of that, I finally added a functionality to the "Reset config" button in the ConfigurationPanel!

I must say all of this is because I have no Internet!!!!