Implemented an API and a library!
This is the first time I did this kind of thing, so I hope it works well.
The main reason I implemented an API it's because I want to show information about the entered game!
Also, I added a library named "json-simple" which will allow me to save these .json responses from the API.
Btw, the name of the API is "RAWG"!
(I will give credits at the developers in a new section of the program!)