Added almost of the "main" functionality of the program
Now, it's possible to add game registers to de program! However, it's not finished yet, because the function of the game register buttons is missing.
Also, I was struggling for almost 3 HOURS because of an error when the program was trying to delete a .json files (which contain data like reviews, screenshots, etc. of the game). Unfortunately, I was no able to solve that problem, so I will continue it in the next days or even at the end of the project...