The main functionality of the program it's (almost) DONE
Now, all the game registers are saved in a file (named "save.dat") when a change is made. This was optional, but I think it must be obligatory, so... because of that, I removed that option, and instead, I added an optional setting that will show a window dialog every time the user wanted to exit.
On the other hand, if the option is enabled, the program will do a backup (only for the registers) before closing the program. This isn't implemented yet, because I want the program to do a backup only if the user made a change, otherwise no. This option is available in the GeneralPanel if the user wants to do a backup anyway.
There are some options that must be implemented before saying that this project is finished, and some of them are:
- Export option
- Help option
- "About" option
- "Wipe out all data" option
Also, there are some ideas I want to do, but these are not very clear or are not totally planned yet.