Removed unused stuff and added new ones!
Everything related to "HTML-table generator" was deleted because I don't see any use further than personal purposes. Also, I deleted everything about my exceptions because at this point they are just an obstacle.
On the other hand, I implemented a "definitively-a-not-good-practice way" to lock files in my program, so now the user (and maybe other processes) won't access/write/delete such files while the program is running. This last change it's an experimental implementation and probably will be deleted if it doesn't work as I want in GNU/Linux distros.