Removed some "garbage" code lines and added more validations
Now, the program verifies if the entire directory "data" exists and creates it if otherwise. However, some validations are not well implemented so probably it will change a bit in the next commits. When a file is going to be saved in a directory, the program will check if that folder exists and then continue operation. If not, first will create it.

Also, I removed a bad implementation I made about the Configuration class.