Added folder and file validations!
Now, the program will "validate and (if necessary) resolve" the directories that are not existing in the same path as the program.
Now, if the user delete some folder, the program just will create it again and use it as it usually does.
Now, if there are not config and save files created, then they will be. Otherwise, the program just will use the existing files with no problem an fear on resetting them.