More configuration options and some changes
Now it's possible to delete all the created backups at the moment, the game info downloaded from the RAWG database, and reset the save.dat file. Also, if you'd like it, you can do all the things just by clicking the "Wipe Out" option.
At this point, there are left just a few things that are not hard to complete, like the "about" and "help" sections. But, I will also comment my code and validate some things about the downloaded info from the database, because there are fields that return null and I need to catch 'em!