Added HelpPanel, so I also added some explanations. And... more stuff
Now, the Help button works! The Help section of the program is not finished yet and probably will change some of the texts because of my lack of English experience. Also, I'm wondering if I have to add some explanations about the "how to create a completed-game register" question. I think a "report bug" button will be great, but I don't know where to put it!