A total mess up, but still working!
Summarizing, I was trying to figure out how to compile my java project in the command line. However, because I spent almost 6 hours (or more) thinking on the same topic, I just gave up and create a new java project in Ecplise with the current sources files. So now, there are a LOT of .class files.
Fortunately, I could make a .jar file that works without the resources folder. This means that, if the user moves the .jar file, then the program will still work and show the images, even if the user didn't download the "res" folder (actually, that was one of the project goals!).
Also, I handled some Exceptions related to these resources, which in case of non-existence, they will be displayed as plain text (I know, it's not the best solution, but note that the images should be visible any moment due the said above).