Updated Advice.getStackTrace function and more little stuff
Now the getStackTrace() function from Advice class shows up the following things:
- The time when the exception occurred
- The return of the "toString()" function
- The class, message, localized message, and cause of the exception
- And the stack trace (maybe will be improved).
On the other hand, I updated Advice.showTextAreaAdvice() functions to request two more parameters "width" (number of columns) and "height" (number of rows) in the JTextArea to make it more flexible depending on the context.
Because of that, I modified some code lines where I call it to make more accurate the size of the JDialog corresponding to the content shown.