Change of JSON library and created pom.xml file
I changed libraries and now I'm using a more recent version of a JSON data management library, so I modified some code lines (fortunately, I wasn't that much).
The reason I've created a pom.xml file was that I want to tell GitHub that I'm using a library from a repository... so, for now, it'll stay there unless I see it useless...

The next step it's adding a "bar" which will contain the text "Recently added registers". This is because when the user creates a new register, the view adds it but not ordered (alphabetically). If I wanted to order the view of those registers, it will be a little be unnecessary. Also, when the user re-enters into the program, it will display the registers ordered alphabetically (due to the GameRegister class).