Added "Recently added" button
This button is just a note to the user telling that the register will be ordered (alphabetically) in the next restart. This is because of the functionality of the program and the way it visually appends registers. At the beginning of the program, they will be ordered because the "add" function in GameRegister class puts the elements alphabetically, but the JPanles can't to it directly unless re-stacking all the inserted panels (imagine doing that with 1000 registers? omg).

Finally, the next step is adding the "About & Credits" section int the program.