Finally, added the AboutPanel class
Not gonna lie, I was struggling quite a pair of hours just trying to figure out how to "properly" append the licenses and terms of use in the program. I mean, yeah, I can put them into a JTextArea but... is it enough? or, is it necessary to do something else? Anyways, I did it in the way I think could be legit-enough so I'm good with it.

Btw, I added a new option to the ViewGamePanel class where the user can go directly to the RAWG's website to see more details about the game displayed in the program.

Now, the next steps are:
- Create a logotype of the project and add it into the About panel with a brief description
- Maybe add some icons in the GeneralPanel class
- Comment the whole project (mmm... maybe at the end)
- Maybe create a website for the project (this will be complicated 'cause of the time, but maybe I will create some sort of "placeholder" and do it later)
- Make a video showcasing the use of the program

I don't know if I will add or remove one of these steps... we will see!