Changed some implementations in GameStat and GameRegister
Now, because of these changes, there's a very good chance to add more GameStat fields in the future.

Previously, the program was saving an ArrayList of GameStats. However, this implies that the fields must be definitive if I wanted to use the SAME save.dat files in future updates, requering a "transform tool" to pass the old file to a new one's format.
Now, the program saves an ArrayList of HashMaps, saving the attributes of the GameStat as a String-Object key-value pair. With this, I can store all the fields I want and I just need adding some additional validations in the classes and, it's done! no more "transform tools"!