Added two more configuration fields and more stuff!
Now the user can change the connect and read timeout in the configuration panel (between 1000 to 10000 milisecs).
Also, I did the same "compatibility" change within the configuration file and the backups. This is because of the same reasons as before, giving more opportunities to add more changes in the configuration file without requering "transforming" the file. In the case of the backup files, I actually forgot to add it, but now it's done.