Modified MainWindow.changePanel() function
Now, every JPanel (except for GeneralPanel) set its JScrollBar at the 0 positions to show the whole page from the beginning.
To achieve this, I added a new JScrollBar variable to (almost) all the JPanels and pass them as parameters to the MainWindow.changePanel function.
However, if the user returns to the "main" panel (GeneralPanel) this will not change its scroll bar position.