Added a new class and updated some stuff
The new class named "Navigation" gives the chance to go to a website page by just callin a function named "goToPage".
So, because of this, some classes have changed in some lines to call this function.

Also, added a new verify function in MainController class used to check if the user has installed the "Open Sans" font, because the program uses it by default.